+++ .gitignore
@@ -151,3 +151,4 @@ dist
 .env*
 cloud-storage-key.json
 database/data
+*.pem
