+++ package.json
@@ -29,6 +29,7 @@
   },
   "dependencies": {
     "@fastify/cors": "^8.1.0",
+    "@fastify/multipart": "^7.1.2",
     "@google-cloud/storage": "^6.4.2",
     "@graphql-tools/merge": "^8.3.6",
     "dotenv": "^16.0.2",
@@ -39,7 +40,6 @@
     "jsonwebtoken": "^8.5.1",
     "lunar-javascript": "^1.2.33",
     "mercurius": "^11.0.0",
-    "multer": "^1.4.5-lts.1",
     "node-fetch": "^3.2.10",
     "pg": "^8.8.0",
     "telegraf": "^4.9.1"
