+++ src/fastify/oauth/sql/awakeBBatonUser.sql
