+++ src/fastify/oauth/sql/awakeBBatonUser.ts
