+++ src/fastify/oauth/sql/registerBBatonUser.sql
