+++ src/fastify/oauth/sql/registerBBatonUser.ts
